pyproject.toml
[project]
name = "10-website-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.2",
    "playwright>=1.57.0",
    "requests>=2.32.5",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2.1",
    "langgraph>=0.4",
    "python-dotenv>=1.1",
]
